backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pynacl==1.5.0
pydantic==2.7.2
websockets==12.0
pyjwt==2.8.0
